+++ src/main/java/de/markusbordihn/easynpc/client/screen/ClientScreens.java
@@ -29,6 +29,7 @@ import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import de.markusbordihn.easynpc.Constants;
 import de.markusbordihn.easynpc.client.screen.configuration.actions.BasicActionConfigurationScreen;
 import de.markusbordihn.easynpc.client.screen.configuration.actions.DialogActionConfigurationScreen;
+import de.markusbordihn.easynpc.client.screen.configuration.actions.DistanceActionConfigurationScreen;
 import de.markusbordihn.easynpc.client.screen.configuration.dialog.BasicDialogConfigurationScreen;
 import de.markusbordihn.easynpc.client.screen.configuration.dialog.NoneDialogConfigurationScreen;
 import de.markusbordihn.easynpc.client.screen.configuration.dialog.YesNoDialogConfigurationScreen;
@@ -72,6 +73,8 @@ public class ClientScreens {
           BasicActionConfigurationScreen::new);
       MenuScreens.register(ModMenuTypes.DIALOG_ACTION_CONFIGURATION_MENU.get(),
           DialogActionConfigurationScreen::new);
+      MenuScreens.register(ModMenuTypes.DISTANCE_ACTION_CONFIGURATION_MENU.get(),
+          DistanceActionConfigurationScreen::new);
 
       // Dialog Configuration Screens
       MenuScreens.register(ModMenuTypes.NONE_DIALOG_CONFIGURATION_MENU.get(),